04_MCP/05_curl/requirements.txt
fastmcp
fast-agent-mcp
docker
